refactored course grouping related pages into folder
