+++ App/src/pages/courseGrouping/DeleteCourseGroupingEdit.tsx
@@ -4,57 +4,44 @@ import {
     Text,
     Heading,
     FormControl,
-    FormLabel,
     Stack,
     Box,
     Textarea,
     useToast,
     FormErrorMessage,
-    Input,
-    TableContainer,
-    Table,
-    Thead,
-    Tr,
-    Td,
-    IconButton,
-    Tbody,
-    Th,
 } from "@chakra-ui/react";
 import { useState, useEffect } from "react";
-import { getAllCourseSettings } from "../services/course";
-import { AllCourseSettings } from "../models/course";
-import { showToast } from "./../utils/toastUtils";
-import Button from "../components/Button";
-import { Link, useNavigate, useParams } from "react-router-dom";
-import { BaseRoutes } from "../constants";
+import { getAllCourseSettings } from "../../services/course";
+import { AllCourseSettings } from "../../models/course";
+import { showToast } from "../../utils/toastUtils";
+import Button from "../../components/Button";
+import { useLocation, useNavigate, useParams } from "react-router-dom";
+import { BaseRoutes } from "../../constants";
 import {
     CourseGroupingDTO,
     CourseGroupingModificationRequestDTO,
     CourseGroupingModificationInputDTO,
-    MultiCourseGroupingDTO,
-} from "../models/courseGrouping";
-import { GetCourseGroupingByName, InitiateCourseGroupingDeletion } from "../services/courseGrouping";
-import { InfoIcon } from "@chakra-ui/icons";
+} from "../../models/courseGrouping";
+import { EditCourseGroupingDeletion } from "../../services/courseGrouping";
 
 export default function DeleteCourse() {
     const { dossierId } = useParams();
     const navigate = useNavigate();
     const toast = useToast();
+    const location = useLocation();
 
     const [isLoading, toggleLoading] = useState(false);
     const [formSubmitted, setFormSubmitted] = useState(false);
-    const [courseGroupingError, setCourseGroupingError] = useState(true);
-    const [rationaleError, setRationaleError] = useState(true);
-    const [resourceImplicationError, setResourceImplicationError] = useState(true);
+    const [rationaleError, setRationaleError] = useState(false);
+    const [resourceImplicationError, setResourceImplicationError] = useState(false);
 
     const [allCourseSettings, setAllCourseSettings] = useState<AllCourseSettings>(null);
-    const [courseGroupings, setCourseGroupings] = useState<CourseGroupingDTO[]>([]);
     const [selectedCourseGrouping, setSelectedCourseGrouping] = useState<CourseGroupingDTO>(null);
 
-    const [searchInput, setSearchInput] = useState<string>("");
     const [rationale, setRationale] = useState("");
     const [resourceImplication, setResourceImplication] = useState("");
     const [comment, setComment] = useState("");
+    const [requestId, setRequestId] = useState("");
 
     const handleChangeRationale = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
         if (e.currentTarget.value.length === 0) setRationaleError(true);
@@ -71,8 +58,13 @@ export default function DeleteCourse() {
     };
     const handleSubmitRequest = () => {
         setFormSubmitted(true);
-        if (courseGroupingError || rationaleError || resourceImplicationError) return;
-        else {
+        if (rationale.trim().length === 0) {
+            setRationaleError(true);
+            return;
+        } else if (resourceImplication.trim().length === 0) {
+            setResourceImplicationError(true);
+            return;
+        } else {
             const subGroupingReferences = selectedCourseGrouping?.subGroupingReferences?.map(
                 // eslint-disable-next-line @typescript-eslint/no-unused-vars
                 ({ id, parentGroupId, ...rest }) => rest
@@ -96,9 +88,9 @@ export default function DeleteCourse() {
                 comment: comment,
                 courseGrouping: courseGroupingRequestDTO,
             };
-            InitiateCourseGroupingDeletion(dossierId, courseGroupingDeletionRequest)
+            EditCourseGroupingDeletion(dossierId, requestId, courseGroupingDeletionRequest)
                 .then(() => {
-                    showToast(toast, "Success!", "Course grouping deletion request successfully created.", "success");
+                    showToast(toast, "Success!", "Course grouping deletion request successfully edited.", "success");
                     toggleLoading(false);
                     navigate(BaseRoutes.DossierDetails.replace(":dossierId", dossierId));
                 })
@@ -114,36 +106,12 @@ export default function DeleteCourse() {
         }
     };
 
-    const searchByName = (name: string) => {
-        GetCourseGroupingByName(name)
-            .then((res: MultiCourseGroupingDTO) => {
-                setCourseGroupings(res.data);
-            })
-            .catch((err) => {
-                if (err.response.status == 400) {
-                    showToast(
-                        toast,
-                        "Error!",
-                        err.response ? "You cannot search an empty string." : "One or more validation errors occurred",
-                        "error"
-                    );
-                }
-            });
-    };
-
-    const handleRowClick = (courseGrouping: CourseGroupingDTO) => {
-        setSelectedCourseGrouping((prevSelectedCourseGrouping: { id: string }) => {
-            if (prevSelectedCourseGrouping && prevSelectedCourseGrouping.id === courseGrouping.id) {
-                setCourseGroupingError(true);
-                return null;
-            } else {
-                setCourseGroupingError(false);
-                return courseGrouping;
-            }
-        });
-    };
-
     useEffect(() => {
+        setRationale(location.state.key.rationale);
+        setResourceImplication(location.state.key.resourceImplication);
+        setComment(location.state.key.comment);
+        setSelectedCourseGrouping(location.state.key.courseGrouping);
+        setRequestId(location.state.key.id);
         getAllCourseSettings()
             .then((res) => {
                 setAllCourseSettings(res.data);
@@ -169,98 +137,12 @@ export default function DeleteCourse() {
                                 >
                                     Back
                                 </Button>
-                                <Stack>
+                                <Stack mt={10} mb={10}>
                                     <Center>
                                         <Heading as="h1" size="2xl" color="brandRed">
-                                            Delete Course Grouping
+                                            Edit Course Grouping Deletion Request
                                         </Heading>
                                     </Center>
-                                </Stack>
-                                <Stack>
-                                    <FormControl isInvalid={courseGroupingError && formSubmitted}>
-                                        <FormLabel htmlFor="search-text">Search Course Grouping By Title:</FormLabel>
-                                        <Input
-                                            id="searcher"
-                                            type="text"
-                                            value={searchInput}
-                                            onChange={(e) => setSearchInput(e.target.value)}
-                                        />
-                                        <FormErrorMessage>You must select a course grouping.</FormErrorMessage>
-                                    </FormControl>
-                                    <Button
-                                        mt={4}
-                                        mb={2}
-                                        style="secondary"
-                                        variant="outline"
-                                        width="22%"
-                                        height="40px"
-                                        onClick={() => {
-                                            searchByName(searchInput);
-                                        }}
-                                    >
-                                        Search
-                                    </Button>
-                                </Stack>
-                                <Stack mb={10}>
-                                    <TableContainer borderRadius="xl" boxShadow="xl" border="2px">
-                                        <Table
-                                            variant="simple"
-                                            style={{ backgroundColor: "white", tableLayout: "auto" }}
-                                        >
-                                            <Thead backgroundColor={"#e2e8f0"}>
-                                                <Tr display={"flex"}>
-                                                    <Th width={"65%"}>Name</Th>
-                                                    <Th width={"15%"}>Credits Required</Th>
-                                                    <Th width={"15%"}>State</Th>
-                                                    <Th width={"15%"} textAlign={"center"}>
-                                                        See Details
-                                                    </Th>
-                                                </Tr>
-                                            </Thead>
-                                            <Tbody>
-                                                {courseGroupings.slice(0, courseGroupings.length).map((grouping) => (
-                                                    <Tr
-                                                        key={grouping.id}
-                                                        display={"flex"}
-                                                        onClick={() => handleRowClick(grouping)}
-                                                        cursor="pointer"
-                                                        background={
-                                                            selectedCourseGrouping &&
-                                                            selectedCourseGrouping.id === grouping.id
-                                                                ? "brandBlue600"
-                                                                : "null"
-                                                        }
-                                                    >
-                                                        <Td width={"65%"} style={{ whiteSpace: "normal" }}>
-                                                            {grouping.name}
-                                                        </Td>
-                                                        <Td width={"15%"}>
-                                                            <Text
-                                                                overflow="hidden"
-                                                                textOverflow="ellipsis"
-                                                                maxW={"500px"}
-                                                            >
-                                                                {grouping.requiredCredits}
-                                                            </Text>
-                                                        </Td>
-                                                        <Td width={"15%"}>{grouping.school}</Td>
-
-                                                        <Td width={"15%"} textAlign={"center"}>
-                                                            <Link
-                                                                to={BaseRoutes.CourseGrouping.replace(
-                                                                    ":courseGroupingId",
-                                                                    grouping.id
-                                                                )}
-                                                                target="_blank"
-                                                            >
-                                                                <IconButton aria-label="Details" icon={<InfoIcon />} />
-                                                            </Link>
-                                                        </Td>
-                                                    </Tr>
-                                                ))}
-                                            </Tbody>
-                                        </Table>
-                                    </TableContainer>
                                 </Stack>
                                 <Stack>
                                     <Center>
